remove redev.h and redev_comm.h circular dep
